ul-30 Updated frontend views and delete registration views
Moved validated registration view function into frontend views module
and deleted old registration view module.
